add image extension condition for command to resize the images

+++ app/Console/Commands/createThumbFromOriginal.php
@@ -80,24 +80,29 @@ class createThumbFromOriginal extends Command
 
         foreach (array_chunk($originalFiles,500) as $key => $value){
             foreach ($value as $k => $orignal) {
-                $bar->advance();
-                // echo Storage::url($orignal)."\n";
-                $fileName = basename($orignal);
-                // echo $fileName."\n";
+                $bar->advance();                
+                $ext = pathinfo($orignal, PATHINFO_EXTENSION);
+                $extension = strtolower($ext);
+                
+                if($extension == 'jpg' || $extension == 'jpeg' || $extension == 'png' || $extension == 'gif'){
+                    // echo Storage::url($orignal)."\n";
+                    $fileName = basename($orignal);
+                    // echo $fileName."\n";
 
-                $pathThumb = public_path($destination .'/'. $fileName);
-                // echo $pathThumb."\n";
-     
-                Image::make($orignal)
-                        ->resize($width, null, function ($constraint) {
-                            $constraint->aspectRatio();
-                        })
-                        ->save($pathThumb);
+                    $pathThumb = public_path($destination .'/'. $fileName);
+                    // echo $pathThumb."\n";
 
-                $thumbImage = $this->fileStorageRepository->addFileToStorage($fileName, $destination.'/', $pathThumb, "s3");                
-                \File::delete($pathThumb);
-                $this->log->info("Completed ".($count)."/".$countAllImages." => ".$fileName);
-                $count++;
+                    Image::make($orignal)
+                            ->resize($width, null, function ($constraint) {
+                                $constraint->aspectRatio();
+                            })
+                            ->save($pathThumb);
+
+                    $thumbImage = $this->fileStorageRepository->addFileToStorage($fileName, $destination.'/', $pathThumb, "s3");                
+                    \File::delete($pathThumb);
+                    $this->log->info("Completed ".($count)."/".$countAllImages." => ".$fileName);
+                    $count++;
+                }
             }
         }
